4주차 미션: 주문 내역 완성
